refactors into hook to consolidate logic
